Move tests into current test project
